refactor: Move iUserSpace functions in Toolkit back to iUserSpace Class
