+++ lib/Deprecated.js
@@ -25,6 +25,7 @@
 /* eslint-disable max-classes-per-file */
 
 const deprecate = require('depd');
+const { parseString } = require('xml2js');
 
 const iPgmDeprecate = deprecate('iPgm');
 const iSqlDeprecate = deprecate('iSql');
@@ -555,29 +556,293 @@ class iProd {
 }
 
 class iUserSpace {
+  /**
+   * @constructor
+   * @param {Connection} conn
+   */
   constructor(conn) {
-    iUserSpaceDeprecate('As of v1.0, class \'iUserSpace\' is deprecated. Please use \'Toolkit\' instead.');
-    this.toolkit = new Toolkit(conn);
+    iUserSpaceDeprecate('As of v1.0, class \'iUserSpace\' is deprecated and will be removed at a later time.');
+    if (conn.constructor.name === 'Connection') {
+      this.conn = conn;
+      this.reportError = conn.returnError;
+    } else if (conn.constructor.name === 'iConn') {
+      this.conn = conn.connection;
+      this.reportError = this.conn.returnError;
+    } else {
+      throw new Error('Expected a Connection object');
+    }
+
+    /* when returnError is false connection.run((xmloutput))
+     * we create a new Connection object so connection.run((error, xmloutput))
+     * this.reportError is checked within the callback to notify the user or not
+    */
+    if (!this.reportError) {
+      this.conn = new Connection({
+        transport: this.conn.transport,
+        transportOptions: this.conn.transportOptions,
+      });
+    }
+
+    this.errno = {
+      type: 'ds',
+      io: 'both',
+      len: 'rec2',
+      fields: [
+        {
+          name: 'bytes_provided',
+          type: '10i0',
+          value: 0,
+          setlen: 'rec2',
+        },
+        { name: 'bytes_available', type: '10i0', value: 0 },
+        { name: 'msgid', type: '7A', value: '' },
+        { type: '1A', value: '' },
+      ],
+    };
   }
 
+  /**
+   * @param {string} name
+   * @param {string} lib
+   * @param {string} attr
+   * @param {number} size
+   * @param {string} auth
+   * @param {string} desc
+   * @param {function} cb
+   */
   createUserSpace(name, lib, attr, size, auth, desc, cb) {
-    iUserSpaceDeprecate('As of v1.0, class \'iUserSpace\' is deprecated. Please use \'Toolkit\' instead.');
-    this.toolkit.createUserSpace(name, lib, attr, size, auth, desc, cb);
+    iUserSpaceDeprecate('As of v1.0, \'iUserSpace.createUserSpace()\' is deprecated and will be removed at a later time.');
+    const pgm = new ProgramCall('QUSCRTUS', { lib: 'QSYS' });
+    const qualifiedName = {
+      type: 'ds',
+      fields: [
+        { type: '10A', value: name },
+        { type: '10A', value: lib === '' ? '*CURLIB' : lib },
+      ],
+    };
+    pgm.addParam(qualifiedName);
+    pgm.addParam({ type: '11A', value: attr === '' ? 'LOG' : attr });
+    pgm.addParam({ type: '10i0', value: size === '' ? 50 : size });
+    pgm.addParam({ type: '1A', value: 0 });
+    pgm.addParam({ type: '10A', value: auth === '' ? '*USE' : auth });
+    pgm.addParam({ type: '50A', value: desc });
+    pgm.addParam({ type: '10A', value: '*NO' });
+    pgm.addParam(this.errno);
+
+    this.conn.add(pgm.toXML());
+    let rtValue; // The returned value.
+
+    const toJson = (transportError, str) => {
+      if (transportError) {
+        if (this.reportError) {
+          cb(transportError, null);
+          return;
+        }
+        cb(null);
+        return;
+      }
+      parseString(str, (parseError, result) => {
+        if (parseError) {
+          if (this.reportError) {
+            cb(parseError, null);
+            return;
+          }
+          cb(str);
+          return;
+        }
+        if (result.myscript.pgm[0].success && result.myscript.pgm[0].success[0].includes('+++ success')) {
+          rtValue = true;
+        } else {
+          rtValue = str;
+        }
+
+        if (this.reportError) {
+          cb(null, rtValue);
+          return;
+        }
+        cb(rtValue);
+      });
+    };
+
+    this.conn.run(toJson); // Post the input XML and get the response.
   }
 
+  /**
+   * @param {string} name
+   * @param {string} lib
+   * @param {string} length
+   * @param {string} msg
+   * @param {function} cb
+   */
   setUserSpaceData(name, lib, length, msg, cb) {
-    iUserSpaceDeprecate('As of v1.0, class \'iUserSpace\' is deprecated. Please use \'Toolkit\' instead.');
-    this.toolkit.setUserSpaceData(name, lib, length, msg, cb);
+    iUserSpaceDeprecate('As of v1.0, \'iUserSpace.setUserSpaceData()\' is deprecated and will be removed at a later time.');
+    const pgm = new ProgramCall('QUSCHGUS', { lib: 'QSYS' });
+    const qualifiedName = {
+      type: 'ds',
+      fields: [
+        { type: '10A', value: name },
+        { type: '10A', value: lib === '' ? '*CURLIB' : lib },
+      ],
+    };
+    pgm.addParam(qualifiedName);
+    pgm.addParam({ type: '10i0', value: 1 });
+    pgm.addParam({ type: '10i0', value: length });
+    pgm.addParam({ type: `${length}A`, value: msg });
+    pgm.addParam({ type: '1A', value: 0 });
+    pgm.addParam(this.errno);
+
+    this.conn.add(pgm.toXML());
+
+    let rtValue; // The returned value.
+
+    const toJson = (transportError, str) => {
+      if (transportError) {
+        if (this.reportError) {
+          cb(transportError, null);
+          return;
+        }
+        cb(null);
+        return;
+      }
+      parseString(str, (parseError, result) => {
+        if (parseError) {
+          if (this.reportError) {
+            cb(parseError, null);
+            return;
+          }
+          cb(str);
+          return;
+        }
+        if (result.myscript.pgm[0].success && result.myscript.pgm[0].success[0].includes('+++ success')) {
+          rtValue = true;
+        } else {
+          rtValue = str;
+        }
+        if (this.reportError) {
+          cb(null, rtValue);
+          return;
+        }
+        cb(rtValue);
+      });
+    };
+
+    this.conn.run(toJson); // Post the input XML and get the response.
   }
 
+  /**
+   * @param {string} name
+   * @param {string} lib
+   * @param {string} length
+   * @param {function} cb
+   */
   getUserSpaceData(name, lib, length, cb) {
-    iUserSpaceDeprecate('As of v1.0, class \'iUserSpace\' is deprecated. Please use \'Toolkit\' instead.');
-    this.toolkit.getUserSpaceData(name, lib, length, cb);
+    iUserSpaceDeprecate('As of v1.0, \'iUserSpace.getUserSpaceData()\' is deprecated ad will be removed at a later time.');
+    const pgm = new ProgramCall('QUSRTVUS', { lib: 'QSYS' });
+    const qualifiedName = {
+      type: 'ds',
+      fields: [
+        { type: '10A', value: name },
+        { type: '10A', value: lib === '' ? '*CURLIB' : lib },
+      ],
+    };
+    pgm.addParam(qualifiedName);
+    pgm.addParam({ type: '10i0', value: 1 });
+    pgm.addParam({ type: '10i0', value: length });
+    pgm.addParam({ type: `${length}A`, io: 'out', value: '' });
+    pgm.addParam(this.errno);
+
+    this.conn.add(pgm.toXML());
+
+    let rtValue; // The returned value.
+
+    const toJson = (transportError, str) => {
+      if (transportError) {
+        if (this.reportError) {
+          cb(transportError, null);
+          return;
+        }
+        cb(null);
+        return;
+      }
+      parseString(str, (parseError, result) => {
+        if (parseError) {
+          if (this.reportError) {
+            cb(parseError, null);
+            return;
+          }
+          cb(str);
+          return;
+        }
+        if (result.myscript.pgm[0].success && result.myscript.pgm[0].success[0].includes('+++ success')) {
+          const { data } = result.myscript.pgm[0].parm[3];
+          rtValue = data[0]._;
+        } else {
+          rtValue = str;
+        }
+        if (this.reportError) {
+          cb(null, rtValue);
+          return;
+        }
+        cb(rtValue);
+      });
+    };
+
+    this.conn.run(toJson); // Post the input XML and get the response.
   }
 
+  /**
+   * @param {string} name
+   * @param {string} lib
+   * @param {function} cb
+   */
   deleteUserSpace(name, lib, cb) {
-    iUserSpaceDeprecate('As of v1.0, class \'iUserSpace\' is deprecated. Please use \'Toolkit\' instead.');
-    this.toolkit.deleteUserSpace(name, lib, cb);
+    iUserSpaceDeprecate('As of v1.0, \'iUserSpace.deleteUserSpace()\' is deprecated and will be removed at a later time');
+    const pgm = new ProgramCall('QUSDLTUS', { lib: 'QSYS' });
+    const qualifiedName = {
+      type: 'ds',
+      fields: [
+        { type: '10A', value: name },
+        { type: '10A', value: lib === '' ? '*CURLIB' : lib },
+      ],
+    };
+    pgm.addParam(qualifiedName);
+    pgm.addParam(this.errno);
+
+    this.conn.add(pgm.toXML());
+
+    let rtValue; // The returned value.
+
+    const toJson = (transportError, str) => {
+      if (transportError) {
+        if (this.reportError) {
+          cb(transportError, null);
+          return;
+        }
+        cb(null);
+        return;
+      }
+      parseString(str, (parseError, result) => {
+        if (parseError) {
+          if (this.reportError) {
+            cb(parseError, null);
+            return;
+          }
+          cb(str);
+          return;
+        }
+        if (result.myscript.pgm[0].success && result.myscript.pgm[0].success[0].includes('+++ success')) {
+          rtValue = true;
+        } else { rtValue = str; }
+
+        if (this.reportError) {
+          cb(null, rtValue);
+          return;
+        }
+        cb(rtValue);
+      });
+    };
+
+    this.conn.run(toJson); // Post the input XML and get the response.
   }
 }
