+++ lib/Toolkit.js
@@ -1747,218 +1747,9 @@ class Toolkit {
   }
 
 
-  createUserSpace(name, lib, attr, size, auth, desc, cb) {
-    const pgm = new ProgramCall('QUSCRTUS', { lib: 'QSYS' });
-    const qualifiedName = {
-      type: 'ds',
-      fields: [
-        { type: '10A', value: name },
-        { type: '10A', value: lib === '' ? '*CURLIB' : lib },
-      ],
-    };
-    pgm.addParam(qualifiedName);
-    pgm.addParam({ type: '11A', value: attr === '' ? 'LOG' : attr });
-    pgm.addParam({ type: '10i0', value: size === '' ? 50 : size });
-    pgm.addParam({ type: '1A', value: 0 });
-    pgm.addParam({ type: '10A', value: auth === '' ? '*USE' : auth });
-    pgm.addParam({ type: '50A', value: desc });
-    pgm.addParam({ type: '10A', value: '*NO' });
-    pgm.addParam(this.errno);
 
-    this.conn.add(pgm.toXML());
-    let rtValue; // The returned value.
 
-    const toJson = (transportError, str) => {
-      if (transportError) {
-        if (this.reportError) {
-          cb(transportError, null);
-          return;
-        }
-        cb(null);
-        return;
-      }
-      parseString(str, (parseError, result) => {
-        if (parseError) {
-          if (this.reportError) {
-            cb(parseError, null);
-            return;
-          }
-          cb(str);
-          return;
-        }
-        if (result.myscript.pgm[0].success && result.myscript.pgm[0].success[0].includes('+++ success')) {
-          rtValue = true;
-        } else {
-          rtValue = str;
-        }
 
-        if (this.reportError) {
-          cb(null, rtValue);
-          return;
-        }
-        cb(rtValue);
-      });
-    };
-
-    this.conn.run(toJson); // Post the input XML and get the response.
-  }
-
-  setUserSpaceData(name, lib, length, msg, cb) {
-    const pgm = new ProgramCall('QUSCHGUS', { lib: 'QSYS' });
-    const qualifiedName = {
-      type: 'ds',
-      fields: [
-        { type: '10A', value: name },
-        { type: '10A', value: lib === '' ? '*CURLIB' : lib },
-      ],
-    };
-    pgm.addParam(qualifiedName);
-    pgm.addParam({ type: '10i0', value: 1 });
-    pgm.addParam({ type: '10i0', value: length });
-    pgm.addParam({ type: `${length}A`, value: msg });
-    pgm.addParam({ type: '1A', value: 0 });
-    pgm.addParam(this.errno);
-
-    this.conn.add(pgm.toXML());
-
-    let rtValue; // The returned value.
-
-    const toJson = (transportError, str) => {
-      if (transportError) {
-        if (this.reportError) {
-          cb(transportError, null);
-          return;
-        }
-        cb(null);
-        return;
-      }
-      parseString(str, (parseError, result) => {
-        if (parseError) {
-          if (this.reportError) {
-            cb(parseError, null);
-            return;
-          }
-          cb(str);
-          return;
-        }
-        if (result.myscript.pgm[0].success && result.myscript.pgm[0].success[0].includes('+++ success')) {
-          rtValue = true;
-        } else {
-          rtValue = str;
-        }
-        if (this.reportError) {
-          cb(null, rtValue);
-          return;
-        }
-        cb(rtValue);
-      });
-    };
-
-    this.conn.run(toJson); // Post the input XML and get the response.
-  }
-
-  getUserSpaceData(name, lib, length, cb) {
-    const pgm = new ProgramCall('QUSRTVUS', { lib: 'QSYS' });
-    const qualifiedName = {
-      type: 'ds',
-      fields: [
-        { type: '10A', value: name },
-        { type: '10A', value: lib === '' ? '*CURLIB' : lib },
-      ],
-    };
-    pgm.addParam(qualifiedName);
-    pgm.addParam({ type: '10i0', value: 1 });
-    pgm.addParam({ type: '10i0', value: length });
-    pgm.addParam({ type: `${length}A`, io: 'out', value: '' });
-    pgm.addParam(this.errno);
-
-    this.conn.add(pgm.toXML());
-
-    let rtValue; // The returned value.
-
-    const toJson = (transportError, str) => {
-      if (transportError) {
-        if (this.reportError) {
-          cb(transportError, null);
-          return;
-        }
-        cb(null);
-        return;
-      }
-      parseString(str, (parseError, result) => {
-        if (parseError) {
-          if (this.reportError) {
-            cb(parseError, null);
-            return;
-          }
-          cb(str);
-          return;
-        }
-        if (result.myscript.pgm[0].success && result.myscript.pgm[0].success[0].includes('+++ success')) {
-          const { data } = result.myscript.pgm[0].parm[3];
-          rtValue = data[0]._;
-        } else {
-          rtValue = str;
-        }
-        if (this.reportError) {
-          cb(null, rtValue);
-          return;
-        }
-        cb(rtValue);
-      });
-    };
-
-    this.conn.run(toJson); // Post the input XML and get the response.
-  }
-
-  deleteUserSpace(name, lib, cb) {
-    const pgm = new ProgramCall('QUSDLTUS', { lib: 'QSYS' });
-    const qualifiedName = {
-      type: 'ds',
-      fields: [
-        { type: '10A', value: name },
-        { type: '10A', value: lib === '' ? '*CURLIB' : lib },
-      ],
-    };
-    pgm.addParam(qualifiedName);
-    pgm.addParam(this.errno);
-
-    this.conn.add(pgm.toXML());
-
-    let rtValue; // The returned value.
-
-    const toJson = (transportError, str) => {
-      if (transportError) {
-        if (this.reportError) {
-          cb(transportError, null);
-          return;
-        }
-        cb(null);
-        return;
-      }
-      parseString(str, (parseError, result) => {
-        if (parseError) {
-          if (this.reportError) {
-            cb(parseError, null);
-            return;
-          }
-          cb(str);
-          return;
-        }
-        if (result.myscript.pgm[0].success && result.myscript.pgm[0].success[0].includes('+++ success')) {
-          rtValue = true;
-        } else { rtValue = str; }
-
-        if (this.reportError) {
-          cb(null, rtValue);
-          return;
-        }
-        cb(rtValue);
-      });
-    };
-
-    this.conn.run(toJson); // Post the input XML and get the response.
-  }
 
   getSysValue(sysValue, cb) {
     let keyValid = false;
